Add QR scanner support for HSAddressInput
