+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/watchaddress/WatchAddressFragment.kt
@@ -9,36 +9,26 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import com.unstoppabledomains.exceptions.ns.NamingServiceException
-import com.unstoppabledomains.resolution.Resolution
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseFragment
-import io.horizontalsystems.bankwallet.core.convertedError
-import io.horizontalsystems.bankwallet.entities.Address
-import io.horizontalsystems.bankwallet.entities.DataState
+import io.horizontalsystems.bankwallet.modules.address.HSAddressInput
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.TranslatableString
 import io.horizontalsystems.bankwallet.ui.compose.components.AppBar
-import io.horizontalsystems.bankwallet.ui.compose.components.FormsInput
 import io.horizontalsystems.bankwallet.ui.compose.components.MenuItem
 import io.horizontalsystems.core.findNavController
-import io.horizontalsystems.ethereumkit.core.AddressValidator
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 class WatchAddressFragment : BaseFragment() {
 
@@ -98,76 +88,12 @@ fun WatchAddressScreen(navController: NavController) {
                 modifier = Modifier
                     .padding(horizontal = 16.dp, vertical = 12.dp)
                     .focusRequester(focusRequester),
-                onValueChange = {
-                    viewModel.onEnterAddress(it)
-                },
-                coinCode = "ETH"
+                coinCode = "ETH",
+                onValueChange = viewModel::onEnterAddress
             )
 
             SideEffect {
                 focusRequester.requestFocus()
-            }
-        }
-    }
-}
-
-@Composable
-fun HSAddressInput(modifier: Modifier, onValueChange: (Address?) -> Unit, coinCode: String) {
-    val viewModel = viewModel<AddressViewModel>(factory = HSAddressInputModule.Factory(coinCode))
-    val scope = rememberCoroutineScope()
-
-    var error by remember { mutableStateOf<String?>(null) }
-
-    FormsInput(
-        modifier = modifier,
-        hint = stringResource(id = R.string.Watch_Address_Hint),
-        error = error,
-        onValueChange = {
-            scope.launch {
-                val addressState = viewModel.parseAddress(it)
-                error = addressState.errorOrNull?.convertedError?.localizedMessage
-                onValueChange.invoke(addressState.dataOrNull)
-            }
-        }
-    )
-}
-
-object HSAddressInputModule {
-    class Factory(private val coinCode: String) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return AddressViewModel(Resolution(), coinCode) as T
-        }
-    }
-}
-
-class AddressViewModel(
-    private val resolution: Resolution,
-    private val coinCode: String
-) : ViewModel() {
-
-    suspend fun parseAddress(v1: String): DataState<Address?> = withContext(Dispatchers.IO) {
-        val v = v1.trim()
-        if (v.length < 40 && !v.contains(".")) {
-            DataState.Success(null)
-        } else {
-            try {
-                var addressString: String = v
-                var addressDomain: String? = null
-
-                if (v.contains(".")) {
-                    if (resolution.isSupported(v)) {
-                        addressString = resolution.getAddress(v, coinCode)
-                        addressDomain = v
-                    }
-                }
-
-                AddressValidator.validate(addressString)
-                DataState.Success(Address(addressString, addressDomain))
-            } catch (e: NamingServiceException) {
-                DataState.Error(e)
-            } catch (e: AddressValidator.AddressValidationException) {
-                DataState.Error(e)
             }
         }
     }
