+++ app/src/main/java/io/horizontalsystems/bankwallet/ui/compose/components/Forms.kt
@@ -1,5 +1,9 @@
 package io.horizontalsystems.bankwallet.ui.compose.components
 
+import android.app.Activity
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
@@ -16,11 +20,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.platform.LocalClipboardManager
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.utils.ModuleField
+import io.horizontalsystems.bankwallet.modules.qrscanner.QRScannerActivity
 import io.horizontalsystems.bankwallet.ui.compose.ColoredTextStyle
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 
@@ -31,6 +38,8 @@ fun FormsInput(
     error: String?,
     onValueChange: (String) -> Unit
 ) {
+    val context = LocalContext.current
+
     val borderColor = when {
         error != null -> ComposeAppTheme.colors.red50
         else -> ComposeAppTheme.colors.steel20
@@ -78,6 +87,15 @@ fun FormsInput(
 
             val clipboardManager = LocalClipboardManager.current
 
+            val qrScannerLauncher = rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+                if (result.resultCode == Activity.RESULT_OK) {
+                    val scannedText = result.data?.getStringExtra(ModuleField.SCAN_ADDRESS) ?: ""
+
+                    textState = textState.copy(text = scannedText, selection = TextRange(scannedText.length))
+                    onValueChange.invoke(scannedText)
+                }
+            }
+
             if (textState.text.isNotEmpty()) {
                 ButtonSecondaryCircle(
                     modifier = Modifier.padding(end = 8.dp),
@@ -88,6 +106,14 @@ fun FormsInput(
                     }
                 )
             } else {
+                ButtonSecondaryCircle(
+                    modifier = Modifier.padding(end = 8.dp),
+                    icon = R.drawable.ic_qr_scan_20,
+                    onClick = {
+                        qrScannerLauncher.launch(QRScannerActivity.getScanQrIntent(context))
+                    }
+                )
+
                 clipboardManager.getText()?.text?.let { textInClipboard ->
                     ButtonSecondaryDefault(
                         modifier = Modifier.padding(end = 8.dp),
@@ -110,5 +136,4 @@ fun FormsInput(
             )
         }
     }
-
 }
